requirements.txt
opencv-python
tqdm
scikit-image
matplotlib
imgaug
memory_profiler
torchvision
segmentation-models-pytorch
pandas
torchmetrics
lightning[extra]
numpy==1.26.4
datasets==3.4.1
scikit-learn==1.6.1
seaborn==0.13.2